RestBlog/requirements.txt
adslproxy==2.0.0
aiohttp==2.1.0
alabaster==0.7.10
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.0
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
APScheduler==3.6.3
asgiref==3.2.7
asn1crypto==0.24.0
astroid==1.5.3
async-timeout==1.2.1
attrs==19.3.0
Automat==0.8.0
Babel==2.5.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.0
bitarray==0.8.1
bleach==2.0.0
boto==2.48.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
conda==4.3.30
conda-build==3.0.27
conda-package-handling==1.6.0
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.9
cssselect==1.1.0
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
decorator==4.1.2
defusedxml==0.5.0
diff-match-patch==20121119
distlib==0.3.0
distributed==1.19.1
Django==2.0.2
django-apscheduler==0.3.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.0
django-debug-toolbar==2.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.5.6
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
docker-py==1.10.6
docker-pycreds==0.4.0
docutils==0.14
drf-extensions==0.3.1
entrypoints==0.2.3
enum-compat==0.0.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.12
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
furl==2.1.0
future==0.16.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
greenlet==0.4.12
gunicorn==19.9.0
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==19.0.0
idna==2.6
imagesize==0.7.1
importlib-metadata==1.6.1
importlib-resources==2.0.1
incremental==17.5.0
ipykernel==4.6.1
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.3
jedi==0.10.2
Jinja2==2.10
joblib==0.15.1
json5==0.8.5
jsonpath-ng==1.4.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
jwt==0.5.3
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
logparser==0.8.2
lxml==4.4.2
Markdown==2.6.11
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.7.4
more-itertools==8.0.2
mpmath==0.19
msgpack-python==0.4.8
multidict==2.1.6
multipledispatch==0.4.9
mysqlclient==1.3.13
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numpy==1.18.4
numpydoc==0.7.0
oauthlib==2.0.7
odfpy==1.3.6
olefile==0.44
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
openpyxl==2.5.2
orderedmultidict==1.0.1
packaging==16.8
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
pep8==1.7.0
pexpect==4.7.0
pickleshare==0.7.4
Pillow==5.0.0
pipenv==2020.6.2
pkginfo==1.4.1
pluggy==0.13.1
ply==3.10
prompt-toolkit==2.0.10
psutil==5.4.0
ptyprocess==0.6.0
py==1.4.34
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
pyee==6.0.0
PyExecJS==1.5.1
pyflakes==1.6.0
pygame==1.9.6
Pygments==2.2.0
PyHamcrest==2.0.0
PyJWT==1.6.1
pylint==1.7.4
pymongo==3.9.0
PyMySQL==0.7.10
pyodbc==4.0.17
pyOpenSSL==19.1.0
pyparsing==2.2.0
pyppeteer==0.0.25
pyquery==1.2.17
PySocks==1.6.7
pyspider==0.3.10
pytest==3.2.1
python-dateutil==2.6.1
python-scrapyd-api==2.1.2
python3-openid==3.1.0
pythonds==1.2.1
pytz==2017.3
PyYAML==5.3.1
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.9.0
queuelib==1.5.0
redis==3.5.3
redis-py-cluster==1.3.6
requests==2.18.4
requests-oauthlib==0.8.0
retrying==1.3.3
rope==0.10.5
ruamel-yaml==0.11.14
schedule==0.5.0
scikit-learn==0.19.2
scipy==1.4.1
scrapinghub==2.3.0
Scrapy==1.5.2
scrapy-crawlera==1.6.0
scrapy-redis==0.6.8
scrapy-redis-bloomfilter==0.7.0
scrapy-redis-cluster==0.5
scrapy-splash==0.7.2
scrapyd==1.2.1
scrapyd-client==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
shub==2.10.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sortedcollections==0.5.3
sortedcontainers==1.5.7
soupsieve==1.9.5
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.3.12
sqlparse==0.3.1
sympy==1.1.1
tablib==0.12.1
tblib==1.3.2
terminado==0.6
testpath==0.3.1
threadpoolctl==2.0.0
toml==0.10.0
toolz==0.8.2
tornado==4.5.2
tqdm==4.40.2
traitlets==4.3.2
tushare==1.2.48
Twisted==19.10.0
typing==3.6.2
tzlocal==2.0.0
u-msgpack-python==2.5.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==20.0.25
virtualenv-clone==0.5.4
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.0.2
wrapt==1.10.11
WsgiDAV==3.0.2
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
xmltodict==0.12.0
yarl==0.10.3
zict==0.1.3
zipp==3.1.0
zope.interface==4.7.1
